CI: Refactor .drone.yaml to use new drone format
This commit is similar to commits
- https://github.com/tolteck/quetzal/commit/21756b1023dd0acb1b151f7c40c1f139741b79b6
- https://github.com/tolteck/coatl/commit/79bf921d30c1a25087088f1967438d43b7a4e339

that were done in Quetzal and Coatl a few years ago: it updates the
format of `.drone.yaml` to take advantage of the newer syntax of
Drone 1.9.

The main advantage is to make maintenance easier by harmonizing our
drone files across our repositories.

This change is a pure refactoring.