CI: Adapt release process for immutable releases
Github recently released a feature that allows for release to be
immutable (meaning that its asset can't be modified once the release
is created).
This feature is slightly better in terms of security but to make use of
it we need to adapt the release process because the github-release
plugin adds the assets after creating the release and this can't work
with immutable releases.
To make it work, let's instead create the release as draft first, and
then turn the release into a 'real' one in a second step.